+++ src/ctx.rs
@@ -2,7 +2,7 @@
 pub struct Context {
     colour: ColourContext,
     stroke: StrokeContext,
-    origin: (f64,f64)
+    origin: (f32,f32)
 }
 #[derive(Debug, Clone, Hash, PartialEq)]
 pub struct ColourContext {
@@ -12,8 +12,8 @@ pub struct ColourContext {
 }
 #[derive(Debug, Clone, Copy, Default, PartialEq)]
 pub struct StrokeContext {
-    inner_strokewidth:f64,
-    outer_strokewidth:f64,
+    inner_strokewidth:f32,
+    outer_strokewidth:f32,
 }
 
 impl Default for ColourContext {
@@ -54,31 +54,31 @@ impl ColourContext {
 
 impl StrokeContext {
     //Might need to revise where I divide by 2
-    pub fn new(strokewidth:f64) -> StrokeContext {
+    pub fn new(strokewidth:f32) -> StrokeContext {
         StrokeContext { 
             inner_strokewidth: strokewidth/2.0,
             outer_strokewidth: strokewidth/2.0,
         }
     }
-    pub fn strokewidth(&self) -> f64 {
+    pub fn strokewidth(&self) -> f32 {
         self.inner_strokewidth + self.outer_strokewidth
     }
-    pub fn i_stroke(&self) -> f64 {
+    pub fn i_stroke(&self) -> f32 {
         self.inner_strokewidth
     }
-    pub fn o_stroke(&self) -> f64 {
+    pub fn o_stroke(&self) -> f32 {
         self.outer_strokewidth
     }
-    pub fn set_i_stroke(&mut self, new_i_stroke:f64) {
+    pub fn set_i_stroke(&mut self, new_i_stroke:f32) {
         self.inner_strokewidth = new_i_stroke;
     }
-    pub fn set_o_stroke(&mut self, new_o_stroke:f64) {
+    pub fn set_o_stroke(&mut self, new_o_stroke:f32) {
         self.outer_strokewidth = new_o_stroke;
     }
 }
 
 impl Context {
-    pub fn new(colour:ColourContext, stroke:StrokeContext, origin:&(f64,f64)) -> Context {
+    pub fn new(colour:ColourContext, stroke:StrokeContext, origin:&(f32,f32)) -> Context {
         Context { colour, stroke, origin:*origin }
     }
     pub fn colour(&self) -> &ColourContext {
@@ -87,13 +87,13 @@ impl Context {
     pub fn stroke(&self) -> &StrokeContext {
         &self.stroke
     }
-    pub fn origin(&self) -> (f64,f64) {
+    pub fn origin(&self) -> (f32,f32) {
         self.origin
     }
-    pub fn new_strokewidth(&self, strokewidth:f64) -> Context {
+    pub fn new_strokewidth(&self, strokewidth:f32) -> Context {
         Context { colour: self.colour.clone(), stroke: StrokeContext::new(strokewidth) , origin:self.origin() }
     }
-    pub fn set_origin(&mut self, svg_origin: (f64, f64)) {
+    pub fn set_origin(&mut self, svg_origin: (f32, f32)) {
         self.origin = svg_origin;
     }
 }
